requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
pymysql==1.0.3
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.7.6
python-docx==0.8.11
